unify error handling
