+++ packages/connect-form/src/components/advanced-options-tabs/advanced-options-tabs.tsx
@@ -8,7 +8,7 @@ import SSHTunnelTab from './ssh-tunnel-tab/ssh-tunnel-tab';
 import TLSTab from './tls-ssl-tab/tls-ssl-tab';
 import AdvancedTab from './advanced-tab/advanced-tab';
 import { UpdateConnectionFormField } from '../../hooks/use-connect-form';
-import { ConnectionFormError } from '../../utils/connect-form-errors';
+import { ConnectionFormError } from '../../utils/validation';
 import { ConnectionOptions } from 'mongodb-data-service';
 
 const tabsStyles = css({
@@ -19,7 +19,6 @@ interface TabObject {
   component: React.FunctionComponent<{
     errors: ConnectionFormError[];
     connectionStringUrl: ConnectionStringUrl;
-    hideError: (errorIndex: number) => void;
     updateConnectionFormField: UpdateConnectionFormField;
     connectionOptions?: ConnectionOptions;
   }>;
@@ -28,13 +27,11 @@ interface TabObject {
 function AdvancedOptionsTabs({
   errors,
   connectionStringUrl,
-  hideError,
   updateConnectionFormField,
   connectionOptions,
 }: {
   errors: ConnectionFormError[];
   connectionStringUrl: ConnectionStringUrl;
-  hideError: (errorIndex: number) => void;
   updateConnectionFormField: UpdateConnectionFormField;
   connectionOptions: ConnectionOptions;
 }): React.ReactElement {
@@ -62,7 +59,6 @@ function AdvancedOptionsTabs({
           <Tab key={idx} name={tabObject.name} aria-label={tabObject.name}>
             <TabComponent
               errors={errors}
-              hideError={hideError}
               connectionStringUrl={connectionStringUrl}
               updateConnectionFormField={updateConnectionFormField}
               connectionOptions={connectionOptions}
