+++ packages/connect-form/src/components/advanced-options-tabs/advanced-tab/advanced-tab.tsx
@@ -15,10 +15,10 @@ import { MongoClientOptions } from 'mongodb';
 
 import FormFieldContainer from '../../form-field-container';
 import { UpdateConnectionFormField } from '../../../hooks/use-connect-form';
-import { ConnectionFormError } from '../../../utils/connect-form-errors';
 import { readPreferences } from '../../../utils/read-preferences';
 
 import UrlOptions from './url-options';
+import { ConnectionFormError } from '../../../utils/validation';
 
 const infoButtonStyles = css({
   verticalAlign: 'middle',
@@ -39,7 +39,6 @@ function AdvancedTab({
 }: {
   errors: ConnectionFormError[];
   connectionStringUrl: ConnectionStringUrl;
-  hideError: (errorIndex: number) => void;
   updateConnectionFormField: UpdateConnectionFormField;
   connectionOptions?: ConnectionOptions;
 }): React.ReactElement {
