+++ packages/connect-form/src/components/advanced-options-tabs/advanced-tab/url-options-table.tsx
@@ -249,7 +249,6 @@ function UrlOptionsTable({
       {errorMessage && (
         <Banner
           variant={'warning'}
-          dismissible={true}
           onClose={() => setErrorMessage('')}
         >
           {errorMessage}
