+++ packages/connect-form/src/components/advanced-options-tabs/authentication-tab/authentication-tab.tsx
@@ -75,7 +75,6 @@ function AuthenticationTab({
 }: {
   errors: ConnectionFormError[];
   connectionStringUrl: ConnectionStringUrl;
-  hideError: (errorIndex: number) => void;
   updateConnectionFormField: UpdateConnectionFormField;
   connectionOptions?: ConnectionOptions;
 }): React.ReactElement {
