--- packages/connect-form/src/constants/markable-form-fields.ts
@@ -1,7 +1,0 @@
-// These are the names of fields which can have errors/warnings shown
-// inline in the form for that field.
-export enum MARKABLE_FORM_FIELD_NAMES {
-  HOSTS = 'HOSTS',
-  IS_SRV = 'IS_SRV',
-  IS_SSH = 'IS_SSH_TUNNEL',
-}
