+++ packages/connect-form/src/hooks/use-connect-form.ts
@@ -4,20 +4,21 @@ import { ConnectionInfo, ConnectionOptions } from 'mongodb-data-service';
 import type { MongoClientOptions } from 'mongodb';
 
 import { defaultConnectionString } from '../constants/default-connection';
-import { ConnectionFormWarning } from '../utils/connect-form-warnings';
-import { ConnectionFormError } from '../utils/connect-form-errors';
+import {
+  ConnectionFormError,
+  ConnectionFormWarning,
+  validateConnectionOptionsWarnings,
+} from '../utils/validation';
 import { getNextHost } from '../utils/get-next-host';
 import { defaultHostname, defaultPort } from '../constants/default-connection';
-import { MARKABLE_FORM_FIELD_NAMES } from '../constants/markable-form-fields';
 import { checkForInvalidCharacterInHost } from '../utils/check-for-invalid-character-in-host';
 import { tryUpdateConnectionStringSchema } from '../utils/connection-string-schema';
 import {
-  handleUpdateConnectionOptions,
-  UpdateConnectionOptions,
-} from '../utils/connection-options-handler';
+  handleUpdateSshOptions,
+  UpdateSshOptions,
+} from '../utils/connection-ssh-handler';
 import { handleUpdateTlsOption } from '../utils/tls-options';
 import { TLS_OPTIONS } from '../constants/ssl-tls-options';
-import { validateConnectionOptionsWarnings } from '../utils/validation-warnings';
 
 export interface ConnectFormState {
   connectionStringInvalidError: string | null;
@@ -40,14 +41,6 @@ type Action =
   | {
       type: 'set-connection-string-state';
       newState: ConnectFormState;
-    }
-  | {
-      type: 'hide-error';
-      errorIndex: number;
-    }
-  | {
-      type: 'hide-warning';
-      warningIndex: number;
     }
   | {
       type: 'set-form-errors';
@@ -75,16 +68,6 @@ function connectFormReducer(
         ...state,
         ...action.newState,
       };
-    case 'hide-error':
-      return {
-        ...state,
-        errors: [...state.errors].splice(action.errorIndex, 1),
-      };
-    case 'hide-warning':
-      return {
-        ...state,
-        warnings: [...state.warnings].splice(action.warningIndex, 1),
-      };
     case 'set-form-errors':
       return {
         ...state,
@@ -95,7 +78,7 @@ function connectFormReducer(
 
 interface UpdateHostAction {
   type: 'update-host';
-  hostIndex: number;
+  fieldIndex: number;
   newHostValue: string;
 }
 
@@ -107,11 +90,11 @@ interface UpdateTlsOptionAction {
 type ConnectionFormFieldActions =
   | {
       type: 'add-new-host';
-      hostIndexToAddAfter: number;
+      fieldIndexToAddAfter: number;
     }
   | {
       type: 'remove-host';
-      hostIndexToRemove: number;
+      fieldIndexToRemove: number;
     }
   | UpdateHostAction
   | {
@@ -122,7 +105,7 @@ type ConnectionFormFieldActions =
       type: 'update-connection-schema';
       isSrv: boolean;
     }
-  | UpdateConnectionOptions
+  | UpdateSshOptions
   | UpdateTlsOptionAction
   | {
       type: 'update-search-param';
@@ -183,7 +166,7 @@ function handleUpdateHost({
   connectionOptions: ConnectionOptions;
   errors: ConnectionFormError[];
 } {
-  const { newHostValue, hostIndex } = action;
+  const { newHostValue, fieldIndex } = action;
   try {
     checkForInvalidCharacterInHost(newHostValue, connectionStringUrl.isSRV);
 
@@ -194,7 +177,7 @@ function handleUpdateHost({
     }
 
     const updatedConnectionString = connectionStringUrl.clone();
-    updatedConnectionString.hosts[hostIndex] = newHostValue || '';
+    updatedConnectionString.hosts[fieldIndex] = newHostValue || '';
 
     // Build a new connection string url to ensure the
     // validity of the update.
@@ -222,8 +205,8 @@ function handleUpdateHost({
       },
       errors: [
         {
-          fieldName: MARKABLE_FORM_FIELD_NAMES.HOSTS,
-          hostIndex,
+          fieldName: 'hosts',
+          fieldIndex,
           message: (err as Error).message,
         },
       ],
@@ -254,14 +237,14 @@ export function handleConnectionFormFieldUpdate({
 
   switch (action.type) {
     case 'add-new-host': {
-      const { hostIndexToAddAfter } = action;
+      const { fieldIndexToAddAfter } = action;
 
       const newHost = getNextHost(
         updatedConnectionStringUrl.hosts,
-        hostIndexToAddAfter
+        fieldIndexToAddAfter
       );
       updatedConnectionStringUrl.hosts.splice(
-        hostIndexToAddAfter + 1,
+        fieldIndexToAddAfter + 1,
         0,
         newHost
       );
@@ -279,9 +262,9 @@ export function handleConnectionFormFieldUpdate({
       };
     }
     case 'remove-host': {
-      const { hostIndexToRemove } = action;
+      const { fieldIndexToRemove } = action;
 
-      updatedConnectionStringUrl.hosts.splice(hostIndexToRemove, 1);
+      updatedConnectionStringUrl.hosts.splice(fieldIndexToRemove, 1);
 
       if (
         updatedConnectionStringUrl.hosts.length === 1 &&
@@ -358,7 +341,7 @@ export function handleConnectionFormFieldUpdate({
           },
           errors: [
             {
-              fieldName: MARKABLE_FORM_FIELD_NAMES.IS_SRV,
+              fieldName: 'isSrv',
               message: `Error updating connection schema: ${
                 (err as Error).message
               }`,
@@ -367,8 +350,8 @@ export function handleConnectionFormFieldUpdate({
         };
       }
     }
-    case 'update-connection-options': {
-      return handleUpdateConnectionOptions(action, {
+    case 'update-ssh-options': {
+      return handleUpdateSshOptions(action, {
         connectionOptions,
         connectionStringUrl,
         errors: initialErrors,
@@ -427,8 +410,6 @@ export function useConnectForm(initialConnectionInfo: ConnectionInfo): [
     setConnectionStringError: (errorMessage: string | null) => void;
     setConnectionStringUrl: (connectionStringUrl: ConnectionStringUrl) => void;
     updateConnectionFormField: UpdateConnectionFormField;
-    hideError: (index: number) => void;
-    hideWarning: (index: number) => void;
     setErrors: (errors: ConnectionFormError[]) => void;
   }
 ] {
@@ -505,18 +486,6 @@ export function useConnectForm(initialConnectionInfo: ConnectionInfo): [
         });
       },
       setConnectionStringUrl,
-      hideError: (errorIndex: number) => {
-        dispatch({
-          type: 'hide-error',
-          errorIndex,
-        });
-      },
-      hideWarning: (warningIndex: number) => {
-        dispatch({
-          type: 'hide-warning',
-          warningIndex,
-        });
-      },
       updateConnectionFormField,
       setErrors: (errors: ConnectionFormError[]) => {
         dispatch({
